rockchip: Enable binman for ARM64
Binman is now being used to build the final flashable images for
Rockchip devices, thus enabling it for all Rockchip targets here. But
it is not yet being used to generate the FIT image (u-boot.itb),
thus we need to force it to be built.

Q: Will this causes issues with eg. Chromebook gru/bob, which build
u-boot.itb with binman already?

Signed-off-by: Andrew Abbott <andrew@mirx.dev>